access for admin added

+++ routes/web.php
@@ -19,10 +19,10 @@ Route::get('/', function () {
 
 Route::group(['namespace' => 'App\Http\Controllers\Weather'], function() {
     Route::get('/weather', 'IndexController')->name('weather.index');
-    Route::get('/weather/create', 'CreateController')->name('weather.create');
-    Route::post('/weather/store', 'StoreController')->name('weather.store');
+    Route::get('/weather/create', 'CreateController')->middleware('access')->name('weather.create');
+    Route::post('/weather/store', 'StoreController')->middleware('access')->name('weather.store');
     Route::get('/weather/{city}', 'ShowController')->name('weather.show');
-    Route::delete('/weather/{city}', 'DestroyController')->name('weather.destroy');
+    Route::delete('/weather/{city}', 'DestroyController')->middleware('access')->name('weather.destroy');
 });
 
 Auth::routes();
